+++ NEWS.md
@@ -1,6 +1,15 @@
 News
 =====
 
+# nestedcv 0.7.4
+###### 30/01/2024
+* Fixed issue with xgboost on linux/windows with parallel processing in 
+`nestcv.train()`. If argument `cv.cores` >1, openMP multithreading is now 
+disabled, which prevents caret models `xgbTree` and `xgbLinear` from crashing, 
+and allows them to be parallelised efficiently over the outer CV loops.
+* Improvements to `var_stability()` and its plots.
+* Fixed major bug in multivariate Gaussian and Cox models in `nestcv.glmnet()`
+
 # nestedcv 0.7.3
 ###### 30/11/2023
 * Added new feature `repeatcv()` to apply repeated nested CV to the main